chore: manager code re-organization (#1314)
This doesn't make any function changes to the manager
it just separates the logic into separate files and
provides some comments for helping to navigate the code.
