+++ railgun/manager/controllerdef.go
@@ -1,0 +1,57 @@
+package manager
+
+import (
+	"fmt"
+	"reflect"
+
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	"github.com/kong/kubernetes-ingress-controller/pkg/util"
+)
+
+// -----------------------------------------------------------------------------
+// Controller Manager - Controller Definition Interfaces
+// -----------------------------------------------------------------------------
+
+// Controller is a Kubernetes controller that can be plugged into Manager.
+type Controller interface {
+	SetupWithManager(ctrl.Manager) error
+}
+
+// AutoHandler decides whether the specific controller shall be enabled (true) or disabled (false).
+type AutoHandler func(client.Reader) bool
+
+// ControllerDef is a specification of a Controller that can be conditionally registered with Manager.
+type ControllerDef struct {
+	IsEnabled   *util.EnablementStatus
+	AutoHandler AutoHandler
+	Controller  Controller
+}
+
+// Name returns a human-readable name of the controller.
+func (c *ControllerDef) Name() string {
+	return reflect.TypeOf(c.Controller).String()
+}
+
+// MaybeSetupWithManager runs SetupWithManager on the controller if its EnablementStatus is either "enabled", or "auto"
+// and AutoHandler says that it should be enabled.
+func (c *ControllerDef) MaybeSetupWithManager(mgr ctrl.Manager) error {
+	switch *c.IsEnabled {
+	case util.EnablementStatusDisabled:
+		return nil
+
+	case util.EnablementStatusAuto:
+		if c.AutoHandler == nil {
+			return fmt.Errorf("'auto' enablement not supported for controller %q", c.Name())
+		}
+
+		if enable := c.AutoHandler(mgr.GetAPIReader()); !enable {
+			return nil
+		}
+		fallthrough
+
+	default: // controller enabled
+		return c.Controller.SetupWithManager(mgr)
+	}
+}
