+++ railgun/manager/flagset.go
@@ -1,17 +1,22 @@
 package manager
 
 import (
-	"github.com/kong/kubernetes-ingress-controller/pkg/util"
 	"github.com/spf13/pflag"
+
+	"github.com/kong/kubernetes-ingress-controller/pkg/util"
 )
+
+// -----------------------------------------------------------------------------
+// Controller Manager - flagSet
+// -----------------------------------------------------------------------------
 
 // flagSet extends flag.FlagSet with additional variable types.
 type flagSet struct {
 	pflag.FlagSet
 }
 
-// EnablementStatusVar defines a flag of type EnablementStatus.
-func (f *flagSet) EnablementStatusVar(p *util.EnablementStatus, name string, value util.EnablementStatus, usage string) {
+// enablementStatusVar defines a flag of type EnablementStatus.
+func (f *flagSet) enablementStatusVar(p *util.EnablementStatus, name string, value util.EnablementStatus, usage string) {
 	*p = value
 	f.Var(p, name, usage)
 }
